Rename qa_tool → sentiment_tool, add langgraph dep, fix memory
- Deleted qa_tool.py — replaced by sentiment_tool.py
- Updated imports in retention_graph.py and retention_agent.py
- Added langgraph to requirements.txt (was missing from PR 48)
- Fixed MemorySaver + thread_id in retention_graph.py
