+++ services/agent-service/requirements.txt
@@ -2,6 +2,7 @@ fastapi==0.115.0
 uvicorn==0.30.0
 langchain==0.3.0
 langchain-aws==0.2.0
+langgraph>=0.2.0
 boto3==1.35.0
 python-dotenv==1.0.0
 httpx==0.27.0
