--- services/agent-service/tools/qa_tool.py
@@ -1,38 +1,0 @@
-# services/agent-service/tools/qa_tool.py
-# LangChain Tool that calls Okino's QA Evaluator SageMaker endpoint
-# George (gvill0576) — Capstone-Churn
-
-import httpx
-import os
-from langchain.tools import tool
-
-QA_URL = os.getenv(key="QA_EVALUATOR_URL", default="http://localhost:8000")
-
-
-@tool
-def analyze_call(transcript: str) -> str:
-    """
-    Analyzes a customer support call transcript for quality and sentiment.
-    Use this tool when given a call transcript to evaluate.
-    Returns a QA score from 0 to 10 and a sentiment classification.
-
-    Args:
-        transcript: The full text of the customer support call
-
-    Returns:
-        JSON string with qa_score and sentiment fields
-    """
-    try:
-        response = httpx.post(
-            url=f"{QA_URL}/predict",
-            json={"text": transcript},
-            timeout=30.0,
-        )
-        response.raise_for_status()
-        return str(object=response.json())
-    except httpx.TimeoutException:
-        return '{"error": "QA service timeout", "qa_score": 5, "sentiment": "unknown"}'
-    except httpx.HTTPStatusError as e:
-        return f'{{"error": "QA service returned {e.response.status_code}", "qa_score": 5, "sentiment": "unknown"}}'
-    except Exception as e:
-        return f'{{"error": "{str(object=e)}", "qa_score": 5, "sentiment": "unknown"}}'
